backend/requirements.txt
fastapi
uvicorn
httpx
PyMuPDF
python-docx
python-multipart
pydantic
pyjwt
youtube-transcript-api
supabase 